tests/requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pyyaml>=6.0
pathlib>=1.0.1
